Implement PowerShell cmdlets (#2838)
This PR implements two new PowerShelll cmdlets Assert-WinGetPackageManager and Repair-WinGetPackageManager as well as the WinGetPackageManager DSC Resource.

Assert-WinGetPackageManager verifies the integrity of winget for the current user. If winget is not set up correctly, it will throw an exception with the type of failure if detected.
It can detect:

Failed call into winget --version.
winget is not found because is not in the PATH environment variable.
winget is not found because the App execution alias is disabled.
Windows OS is not supported.
The AppInstaller package is not installed in the machine.
The AppInstaller package is not registered for the user.
The AppInstaller package is old and doesn't contain winget.
Assert-WinGetPackageManager -Version foo also verifies the installed version is the same as the expected.
Assert-WinGetPackageManager -Latest also verifies the installed version is the latest release version.
Assert-WinGetPackageManager -Latest -IncludePrerelease also verifies the installed version is the latest prerelease version.

Repair-WinGetPackageManager does the same as Assert-WinGetIntegrity but attempts to fix winget depending on the failure. Returns 0 is succeeded.

Repair-WinGetPackageManager repairs current winget installed and performs no updates. If winget is not installed or there's catastrophic failure then installs the latest released winget.
Repair-WinGetPackageManager -Version foo repairs winget and makes sure that the current installed version is the expected one.
Repair-WinGetPackageManager -Latest repairs winget and makes sure winget is the latest released version.
Repair-WinGetPackageManager -Latest -IncludePreRelease repairs winget and makes sure winget is the latest prereleased version.

It sadly can't repair if AppExecution alias is disabled because there's no programmatic way to do it (requires internal APIs).

It also doesn't install Microsoft.UI.Xaml.2.7 as a dependency because the current solution to download the nuget package, extract it and install the appx within is not the right approach we want to follow. We will follow up internally with the owners of the package to do something similar as the VCLibs packages.

There's still a lot of opportunity for Repair-WinGetPackageManager to make it more robust, but this PR is just the starting phase.

The WinGetPackageManager DSC resource is just a simple wrapper around those two cmdlets. There's also a sample on how to use it.

For now, all versions need to be the tag name of the releases in GitHub.

Also, moved Get-WinGetVersion from Crescendo to a binary cmdlet.

---------

Co-authored-by: JohnMcPMS <johnmcp@microsoft.com>