Make ConfigurationRemotingServer selfcontained (#3097)
This PR removes the dependency on the .net6 runtime for ConfigurationRemotingServer by making it selfcontained. The downside is that now we carry it in the package, but oh well...

I had a long fight with PowerShell's contentFiles to pick the right modules, so what I ended up doing is to exclude contentFiles and add them manually as an AfterBuild task. We need the ref directory and the win modules in the same location as System.Management.Automation.dll