Put required modules into the repo and consume them in dev scenarios (#3103)
Adds the newer version of `PowerShellGet` (and its dependencies) that isn't installed by default on Windows so that a clean machine can build and run the dev package and tests.  Changes the tests and dev package to consume the modules as well.

This change also updates the CI pipeline to produce a full package layout for testing.  This will enable the full package experience to be tested in the end-to-end tests.