Make entitlement best effort (#3172)
Make entitlement acquisition best effort rather than required. A recent change caused some issues for us regarding the entitlement process, and since the entitlement isn't actually required to install, just attempt to get it and log if it failed.  Even if the underlying issue is resolved, this change will prevent future issues from having the same level of impact.

This may lead to an issue where the entitlement acquisition fails, the package is installed, but attempting to run it fails because the entitlement cannot be acquired then either (by the OS in the background).  Given the choice between the two failure modes though, it seems far better to remove these false negatives and create the (my gut says fewer) false positives that will result.