Allow prerelease modules for configuration (#3217)
This PR allows a configuration file to have a boolean allowPrerelease directive to use prerelease DSC resources.

```
properties:
  configurationVersion: 0.2
  resources:
    - resource: FakeModule/FakeResource
      directives:
        version: '1.0.0-preview'
        allowPrerelease: true
      settings:
        Args: 'args'
```

Internally, allow prerelease is used only for a module that it needs to get install. When is enabled, the processor will use `-AllowPrerelease` to the Find-DscResource command. 

It also fixes and infinite loop in `SplitIntoLines`. If the found offset is the one of "\r\n", it uses the same offset for `pos` in `find_first_of`, so it keeps returning the offset making `nextOffset - currentOffset = 0`