Queue write operations for Start-* cmdlets and hook up diagnostics (#3222)
When a Start-* cmdlet starts its operation it won't write to any PowerShell stream because control is returned back to the user almost immediately. This changes queue messages to be then displayed at Continue-* time. This way, we will have all the messages in the appropriate order. It applies for WriteDebug, WriteVerbose, WriteWarning, WriteError and WriteProgress.

WriteProgress will only be called when the activity hasn't been completed. A caller must use AsyncCommand.GetNewProgressActivityId to generate a new activity id.

Also enables telemetry and OnDiagnostic events