Microsoft.WinGet.Configuration messages (#3242)
This PR starts printing information messages for Get-WingetConfiguration and Get-WingetConfigurationDetails. The messages are similar to what winget is showing.

Also:

Add resource file with necessary configuration messages. These are mostly the same as the one winget uses.
Modify AsyncCommand to always use a BlockingCollection for storing messages. The messages are written to PowerShell when a cmdlet waits for the task to complete or a Complete-* cmdlet executes and the task is already completed. This completely removes the context switch (which might come back if we ever require a ShouldContinue/ShouldProgress in the middle of the task)