Configuration cancellation support (#3244)
Adds cancellation support to the configuration `Async` functions, as well as implementing most of the non-async functions without using an extra thread.  The goal is that all cancellations will result in the same `HRESULT` being returned, regardless of which code is actually detecting the cancellation and ending early.

This change also updates the `winget` configuration workflows to pass along cancellation requests and better respond when things are cancelled.

Finally, using the previously implemented progress message system, cancellation will generically inject a message to the user so that it is clear that a `CTRL+C` signal has been received.