Stub upgrade self (#3299)
This is in preparation of the new AppInstaller bundle that will contain stub packages for all the application packages. The stub package doesn't contain any of the configuration binaries and it will be the preferred package installed inbox on Windows.

To install the main package winget configuration --enable must be called. This will detect which type of package is installed and install the full package if needed. winget configuration --disable will install the stub package.

By design we are not supporting winget configuration --file config.yaml --enable because winget itself is updating. Running the same command automatically after the update can only be done by starting a new shell window losing context of the current shell and making scripts hard to implement.

The next PR will be to handle gracefully when deployment is telling us to shut down.

Based on JohnMcPMS/stub-upgrade-self

Co-authored-by: JohnMcPMS <johnmcp@microsoft.com>