Support for out of process configuration clients (#3363)
The goal of this rather large change is to support out-of-process clients to configuration.  This enables supported scenarios to not require shipping the entire PowerShell infrastructure, and instead leverage the one that is contained in our package.

The interface changes are breaking, but minimal.  The goal is to remove the necessity for a processor to need the configuration module itself, while creating a single entry point that can be used for OOP cases (and many in-proc cases as well).