Support winget installing AppInstaller (#3377)
When winget is updating AppInstaller (aka. itself) via winget install 9NBLGGH4NNS1 -s msstore --force if appears the installation failed but in reality, it got updated successfully.

winget install 9NBLGGH4NNS1 -s msstore --force
Verifying/Requesting package acquisition...
Starting package install...
  ███████████████████████████     90%
Failed to install or upgrade Microsoft Store package. Error code: 0x80073d02
The failure means ERROR_PACKAGES_IN_USE and from my last PR #3299 I believe that error will go away because of --force, but a new one will occur. I need a real AppInstaller package to confirm this but because reasons I cannot build it right now. Regardless, the correct fix it to listen to messages from the lifetime manager.

When the system is updating an app, the app lifetime manager will ask any the app to terminate if its running. It does it by:

Sending a WM_QUERYENDSESSION message to any app window.
Sending a CTRL-C signal.
This means that winget.exe needs to start listening to window messages because there's no real difference between the app lifetime manager and a user doing CTRL-C. This PR creates a hidden window and start the message loop. When WM_QUERYENDSESSION is received, we will ask to cancel any progress (as it currently does with CTRL-C) stating the reason as AppShutdown. By design, we are not going to cancel installations as maybe they get completed before they terminate us. Any other operation will be cancelled.

If the context gets cancelled with an AppShutdown reason while installing AppInstaller we will just set the progress as 100% and be done with it as we are going to get terminated soon.