configure test command "implemented" (#3414)
Implemented the `configure test` command.  This was just hooking up the command to a workflow that largely mimics the apply and refactoring the progress handler to share code between the two.  The actual "test the state of a configuration set" function in the configuration module was implemented a while ago.