Simplify creating local index (#3445)
The goal of this PR is to simplify how the test source generated and have a convenient way recreate it locally or to be consume that other agents (aka PowerShell E2E).

There's a new `WinGetSourceCreator` project that incorporates what the E2E setup and IndexCreationTool did. `WinGetLocalSource.CreateLocalSource` is a convenience method in charge of processing that object by:

1. Preparing all local installers by copying them to the working directory and signing them if needed.
2. Preparing all dynamic installers. A dynamic installer is on that is generated while creating the source. Currently it only supports zip installers.
3. Copying all manifest files into the working directory and replace any hash token from previously processed installers. Includes signature hash for msix.
4. Create the index.db file using WinGetUtil
5. Create source.msix and sign the package

`LocalSource` is an object that declares all the artifacts needed to produce a local winget source and is the input of `WinGetLocalSource.CreateLocalSource`. It can also be serialized and consumed as a json file by `WinGetLocalSource.CreateFromLocalSourceFile`. The input json can also reference environment variables (in its Windows version for batch) and they will be expanded by `WinGetLocalSource.CreateFromLocalSourceFile`. 

The E2E local source json is the following:
```
{
  "AppxManifest": "%BUILD_SOURCESDIRECTORY%/src/AppInstallerCLIE2ETests/TestData/Package/AppxManifest.xml",
  "WorkingDirectory": "%AGENT_TEMPDIRECTORY%/TestLocalIndex",
  "LocalManifests": [
    "%BUILD_SOURCESDIRECTORY%/src/AppInstallerCLIE2ETests/TestData/Manifests"
  ],
  "LocalInstallers": [
    {
      "Type": "exe",
      "Name": "AppInstallerTestExeInstaller/AppInstallerTestExeInstaller.exe",
      "Input": "%BUILDOUTDIR%/AppInstallerTestExeInstaller/AppInstallerTestExeInstaller.exe",
      "HashToken": "<EXEHASH>"
    },
    {
      "Type": "msi",
      "Name": "AppInstallerTestMsiInstaller/AppInstallerTestMsiInstaller.msi",
      "Input": "%BUILD_SOURCESDIRECTORY%/src/AppInstallerCLIE2ETests/TestData/AppInstallerTestMsiInstaller.msi",
      "HashToken": "<MSIHASH>"
    },
    {
      "Type": "msix",
      "Name": "AppInstallerTestMsixInstaller/AppInstallerTestMsixInstaller.msix",
      "Input": "%BUILD_ARTIFACTSTAGINGDIRECTORY%/AppInstallerTestMsixInstaller.msix",
      "HashToken": "<MSIXHASH>",
      "SignatureToken": "<SIGNATUREHASH>"
    }
  ],
  "DynamicInstallers": [
    {
      "Type": "zip",
      "Name": "AppInstallerTestZipInstaller/AppInstallerTestZipInstaller.zip",
      "Input": [
        "%AGENT_TEMPDIRECTORY%/TestLocalIndex/AppInstallerTestExeInstaller/AppInstallerTestExeInstaller.exe",
        "%AGENT_TEMPDIRECTORY%/TestLocalIndex/AppInstallerTestMsiInstaller/AppInstallerTestMsiInstaller.msi",
        "%AGENT_TEMPDIRECTORY%/TestLocalIndex/AppInstallerTestMsixInstaller/AppInstallerTestMsixInstaller.msix"
      ],
      "HashToken": "<ZIPHASH>"
    }
  ],
  "Signature": {
    "CertFile": "%APPINSTALLERTEST_SECUREFILEPATH%"
  }
}
```
`src/AppInstallerCLIE2ETests/README.md` contains more information about how to run the tests

The source is generated but in order to be usable by winget it needs to be hosted in the localserver web host.

Before this change, each run of the E2E will generate the same test source. Now it will be produced at the Launch LocalhostWebServer time and the E2E assumes everything is set up correctly. The E2E can still generate the source given the necessary arguments if needed. The code that generates the source is the same. 

I also modified how the test runtime parameters are consumed in the E2E project. This should make it easier to anyone to run them locally without that much pain for test that don't require the test source or the localhost web server. I also moved how the com traces can be enabled via the System.Debug variable so it can work from master too.

Verified by having the same number of test passed/skip in each phase.

```
Run E2E Tests Packaged
Total tests: 222
    Passed: 215
    Skipped: 7

Run Microsoft.Management.Deployment E2E Tests (In-process)
Total tests: 57
    Passed: 52
    Skipped: 5

Run Microsoft.Management.Deployment E2E Tests (Out-of-process)
Total tests: 57
    Passed: 53
    Skipped: 4
```

 ###### Microsoft Reviewers: codeflow:open?pullrequest=https://github.com/microsoft/winget-cli/pull/3445&drop=dogfoodAlpha