Move functions to cmdlets for Microsoft.WinGet.Client (#3469)
This PR moves the Crescendo generated functions to cmdlets for the Microsoft.WinGet.Client module. This is to have a centralized project that manages all the cmdlets and facilitate E2E.

For example, when the client module is built wingetdev was not used in the functions because winget.exe was hardcoded by the Crescendo tool. To properly test it, Initialize-LocalWinGetModules.ps1 find and replaced winget.exe with wingetdev

Funtions to Cmdlets moved:
- Add-WinGetSource
- Disable-WinGetSetting
- Enable-WinGetSetting
- Get-WinGetSettings
- Remove-WinGetSource
- Reset-WinGetSource

I added a `-AsPlainText` Get-WinGetSettings. If enabled it will just print the json string from `winget settings export`, if disabled it will return a HashTable. I also added running as admin check in all the cmdlets that needed and throw an not supported exception if these cmdlets are run in a system context.