Move Microsoft.WinGet.Client E2E test to Pester framework (#3503)
This PR finally uses all the refactor to run the Microsoft.WinGet.Client PowerShell module in Pester. The tests are done in the BuildPowerShellModule phase.

The E2E Test Source is built in the build phases, then it will be get published as artifacts under E2ETests\TestLocalIndex'. Other artifacts that are required as published in E2ETests` as well. Having the test source as a build artifacts will make it easier to perform repros without building the source all over again.

The BuildPowerShellModule phase will download the artifacts, move necessary files to around, install the dev package (and dependencies), start the localhost web server and run the Pester tests. Then it will publish the modules as it used to as well as the test results.

It is important to remember the modules published as artifacts target the dev package, so one cannot just download it and use it without it.

I port all the E2E tests to Pester tests except the ones dealing with the server shutdown.

Eventually we will run tests for Microsoft.WinGet.DSC and Microsoft.WinGet.Configuration.