Remove CrossProcessReaderWriteLock (#3549)
Remove `CrossProcessReaderWriteLock`, as it has caused more problems than it likely ever solved.

The current issue is that the mutexes can potentially be released on a different thread than they were acquired in the COM server.  This led to initially removing them from being stored for the entire lifetime of the index file, after confirming that even in immutable mode, SQLite is opening the file and *not* sharing delete.

After that, the massive complexity of the code is unnecessary, since we ever only hold the lock shared to open the index.  So rather than keep it, I swapped out all usage for the simple wrapper around a named mutex.  I also added a `FAIL_FAST_IF` the current thread was not the one that we had originally acquired the lock on.

This might lead to slightly more waiting while other callers are opening files, but it might just as easily lead to less waiting due to the single mutex.