Enable configuration via IConfigurationStatics (#3576)
Winget configuration is not going to be enabled by default. The user needs to opt-in by calling winget configuration --enable. Callers that are interested in using the OOP Microsoft.Management.Configuration APIs need a way to ensure configuration is enabled and enable it if not. This PR address that.

New IConfigurationStatics methods
IsConfigurationEnabled - Checks if the configuration is enabled by checking if the current AppInstaller package is the stub or full package. Inproc implementation always returns true.
EnableConfigurationAsync - Asynchronously enables configuration. This changes the stub preference to full package and calls our own APIs to install AppIntaller. By design, this also updates AppInstaller to the newest version.
Caveats
Callers need to understand that this is effectively saying "winget please update yourself". Which means that once EnableConfigurationAsync is done, configuration isn't ready yet. At this point, deployment scheduled AppInstaller for an update. The server process that owns the configuration statics object is already gone (or is about to) and a new instance of ConfigurationStaticFunctions needs to be created in order validate configuration is enabled. It is possible that creating that object fails with an error that the upgrade is in progress. Yes, this is awkward.

Testing
It is complicated to write automated tests for this without faking most of the stack. I manually created an AppInstaller package with an older version to what we internally have to do these tests. Then I used the current OOP configuration tests to update it.