SourceAutoUpdateInterval - Group Policy Description update.
[WHY:]
The previously use GPO description for "Set App Installer Source Auto Update Interval In Minutes" was not clear and it was often confused with auto-update of application in the connected sources.

[WHAT:]
The GPO Policy description for "Set App Installer Source Auto Update Interval In Minutes" has been updated to make it more clear.

[How Validated:]
- Patched updated AMDL (adn AMDX) on a test machine
- Launched gpedit
- Confirmed the updated description in reflecting on the GPO