Rework the error list and add a command to get error information (#3615)
This change reworks the error description switch statement into a sorted array of data. We then use binary search to find the appropriate description when needed. In addition, it now supports these values being localized.

This change also adds the `error` command (hidden) to enable a somewhat easier path to getting error information specific to winget.  The command takes in a single value.  If this is wholly a number, it is treated as the error number itself.  If it is not a number, it is treated as a string to search within the symbols and descriptions of the winget errors.