Make manifest retrieval choice more dynamic (#3738)
This change lets the source implementation indicate whether manifests might contain more information than the search results.  It then only applies that to REST, as the implementation is not under our control.  This should have no net effect on the results quality, but should not attempt to retrieve manifests from the winget source when they are not providing any additional value currently.