Use package version as potential last update timestamp (#3759)
This change moves the background update determination to be under the control of the source reference, and then uses the package version (which is based on the creation time for us) as a potential later "last update time".  This enables an update to the package from outside of our control to still prevent any update check from occurring.