Inform user if a module requires elevation (#3758)
Fix #3747

Add a new HRESULT for failed importing module because the module requires to run as admin.
Also fix places where WINGET_CONFIG_ERROR_UNIT_SETTING_CONFIG_ROOT was missing.
We can't have automated tests because we don't currently have a way to run winget as non admin in our tests. I created a UT with a simple test that is skipped in ADO.
I also removed printing details for WINGET_CONFIG_ERROR_UNIT_IMPORT_MODULE because I noticed that "error code could not be found" message is coming the IConfigurationUnitResultInformation description.

Apply :: VSComponents
  Loading the module for the configuration unit failed.
The text associated with this error code could not be found.
Some of the configuration was not applied successfully.
In the configuration module, the IConfigurationUnitResultInformation description is the actual error message that we set in ImportModuleException. My guess is that cppwinrt use some method to get message of well-known errors and set that, but I didn't want to go that deep, so removing it because we are literally printing a line before the text associated with that error code.