Fix pipeline build error (#3937)
This fixes the recent build errors in the pipelines. The compiler was
complaining about re-throwing an exception with `throw e;` instead of
`throw;`, which messes with the stack trace. Presumably it started
because of a tool update in the agents.

This also fixes a compilation error I'm seeing locally from a member not
being initialized.